fix: fix whitelisting rules breaking
This bug was caused by the updating indexing UI not using the same hook as the whitelistIndicator and the content script.
To fix it I have updated the UI to use the same hook as the whitelistIndicator and the content script.
And to prevent this issue happening again I have refactored the useIndexMatching hook to be more generic and reusable.