fix: fix background script error
The background script ran handleIndexPage as if it was an async function, but it wasn't.
This caused the background script to throw an error when it tried to call the next function in the chain.
This commit fixes the error by removing the then() call from handleIndexPage in favour of a try/catch block .