fix: fix message unreceived error
Previously if the service worker became inactive because chrome terminates it after 30s and the user interacted with the sidebar
the sidebar would send a message to the service worker but the service worker would not receive it because it was inactive.
This was fixed by moving the code that registers the message listener to the top of the file so that it is executed when the service worker is activated.

Closes #74