feat: refactor todo list to use single source of truth
This commit todolist logic that was previously in TaskSection and moved it to the useTodoList custom hook.
To accomplish this, the chrome storage setting calls that were split between the useTodoList hook and the background script were consolidated into the background script.
This change was made to make the background script the single source of truth for the todo list data.

the logic that was previously in the TaskSection component concerned recovering completed and deleted tasks.
The logic for completing tasks now works like this
checking a task sets the completed property in the task object to true
moving tabs (unloading the todolist component) will remove all the completed tasks from the tasks array and store them in the completedTasks array